Deal with unauthenticated git from github
https://github.blog/2021-09-01-improving-git-protocol-security-github/

This has announced that as of march 15 2022, github would no longer
support the `git://github.com/user/repo` format. However the
`git@github.com:user/repo` format does work.

This updates the one test I could fix to that. `conflicting_erlydtl` is
dropped because it relies on all applications that transitively also had
the unsupported protocol and that use case is just not exercisable
anymore.